Database and json data.
This zip contains all the data that needed to be read in and the code to read the json data. Since it was given in separate files and there is the possibility to add additional json files, the loading of the json files is tedious but necessary given the constraints of the data I was given.